completed products dummy data

+++ Models/Product.cs
@@ -13,7 +13,7 @@ namespace ERPProject.Models
         public string Name { get; set; }
         public string Category { get; set; }
         public string Description { get; set; }
-        public float CurrentPrice { get; set; }
+        public double CurrentPrice { get; set; }
         public int Stock { get; set; }
         public int Purchases { get; set; }
     }
